Part 1 : Manual dependency injection
Construct the objects a car needs (engine, wheels, etc) from outside the Car class, and then pass them to the Car object at creation time. Do all of this in MainActivity.

This is valid dependency injection because the car does not need to make these objects on its own, but it will eventually require a lot of boilerplate code to be present in MainActivity.

Dagger2 can fix this by creating all of the car parts by itself. This will get rid of the boilerplate code in MainActivity, while still reaping the benefits of dependency injection.