Part 2 : Dagger dependency injection (1)
Dagger2 can automate the creation of objects. Instead of manually creating objects using code, we can instead rely on Dagger to create them for us, so long as all pieces are hooked up to Dagger correctly.

The most important part of Dagger is the Component. The component is responsible for generating objects of the specified type. If we annotate the constructor of the Car class and the constructors of each object it depends on, Dagger can automatically create the Car and return it to us.

Now MainActivity is much simpler. It just needs to create the component and then ask the component for the things we need.