Part 3 : Dagger Field Injection
So far we've learned that the Component can provide new instances of objects through provider methods (see Part 2). A component can have many provider methods, and you can ask the component for objects whenever you need them.

Alternatively, a class can be injected into the Component and have all of its @Inject-marked fields auto populated with the necessary objects. This is called field injection.

With field injection, the Component will be cleaner and all you have to do in the Activity is mark each of the fields you need with @Inject. This ultimately means less code.