Part 6 : Providing Interfaces
To use interfaces with Dagger, we need to tell Dagger which implementations it should use with modules.

Here we make Engine an interface and create a PetrolEngine and a DieselEngine to implement it. We make modules for each and then specify which module to use in the CarComponent list. We can specify only one of these in the component, or else Dagger will not know which module to use to satisfy a request for an Engine.

To make the modules more efficient, we can make them abstract and then use abstract Bind methods, which take in either a PetrolEngine or DieselEngine as parameters, and return an Engine. This makes it so that Dagger doesn't even have to create the Module classes and can just access the methods directly.