Part 7 : Inject Values at Runtime
When an @Inject-ed object needs a value passed to its constructor at runtime, we can pass the value we need (in this case horsePower) to its module when the Component is created. This module can then use that default value for horsePower to manually create DieselEngines that have that horsePower.

Question : Does this assume that all DieselEngines I want provided will have the same horsePower? Looks like it... I guess I can use setters if I ever want to change it.