Part 8 : Component.Builder, BindsInstance, and Named
We can inject values that our objects' constructors need into a Component.Builder instead of doing it through a Module, which requires an instance of the module to be made. This requires that we make our own Component.Builder and provide binder methods for the types we need.

The methods should be annotated with @BindsInstance and return a Builder. The single parameters they expect are the most important part, as the parameter type is what the binding is for. If we have multiple binder methods, we can use an @Named annotation to differentiate between them.

Once all of the is done in the Builder, the module can go back to being a simple abstract class.