Part 11 : Scopes and Component Dependencies
A single object can be created per component by using scopes. Each component is separate, so even if you use the same scope name in another Component it will have a different instance.

Components can live in custom scopes. The scope for objects created in the Application Component might be different from the scope used for the Activity Component. Objects created in the application Component will be true singletons and will live as long as the application is alive, while objects created in the activity scope will only live as long as the activity is alive.